style interface cleanup
