+++ src/web/index.html
@@ -6,21 +6,6 @@
     <script type="module" src="./app.tsx"></script>
     <link rel="stylesheet" href="https://unpkg.com/mini.css" />
     <link rel="stylesheet" href="./web-styles.css" />
-    <style>
-      .tts-audio-visualizer {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        width: 10rem;
-        height: 10rem;
-      }
-
-      .tts-audio-visualizer-bar {
-        background-color: black;
-        border-radius: 10px;
-        width: 6%;
-      }
-    </style>
   </head>
   <body></body>
 </html>
